Move sendAllApps to AppManagerCacheValidatorReceiver.kt
