+++ appmanager/wear/src/androidMain/kotlin/com/boswelja/smartwatchextensions/appmanager/AppManagerCacheValidatorReceiver.kt
@@ -1,7 +1,9 @@
 package com.boswelja.smartwatchextensions.appmanager
 
 import android.content.Context
+import com.boswelja.watchconnection.common.message.Message
 import com.boswelja.watchconnection.common.message.ReceivedMessage
+import com.boswelja.watchconnection.serialization.MessageHandler
 import com.boswelja.watchconnection.serialization.MessageReceiver
 import com.boswelja.watchconnection.wear.discovery.DiscoveryClient
 import com.boswelja.watchconnection.wear.message.MessageClient
@@ -26,7 +28,43 @@ class AppManagerCacheValidatorReceiver :
         // Get the hash code for our local app list, and check against the remote cache
         val currentHash = CacheValidation.getHashCode(currentPackages)
         if (message.data != currentHash) {
-            context.sendAllApps(messageClient, discoveryClient)
+            sendAllApps(context)
         }
     }
+
+    /**
+     * Send all apps installed to the companion phone with a given ID.
+     * @param context [Context].
+     */
+    private suspend fun sendAllApps(
+        context: Context
+    ) {
+        val pairedPhone = discoveryClient.pairedPhone()!!
+
+        val messageHandler = MessageHandler(AppListSerializer, messageClient)
+        // Get all current packages
+        val allApps = context.getAllApps()
+
+        // Let the phone know what we're doing
+        messageClient.sendMessage(
+            pairedPhone.uid,
+            Message(APP_SENDING_START, null)
+        )
+
+        // Send all apps
+        messageHandler.sendMessage(
+            pairedPhone.uid,
+            Message(
+                APP_LIST,
+                AppList(allApps)
+            )
+        )
+
+        // Send a message notifying the phone of a successful operation
+        messageClient.sendMessage(
+            pairedPhone.uid,
+            Message(APP_SENDING_COMPLETE, null)
+        )
+    }
+
 }
